requirements.txt
pinecone
pypdf
langchain
langchain_openai
langchain-community
python-dotenv
tqdm
openai
streamlit
